\name{mosaicMenuItem}
\alias{mosaicMenuItem}
\alias{assocMenuItem}
\alias{weHaveMosaicableTables}
\docType{package}
\title{Rcmdr menu items for mosaic and assoc plots}
\description{Rcmdr menu items for mosaic and assoc plots}
\usage{
mosaicMenuItem()
weHaveMosaicableTables()
}


\note{Special thanks go to Richard Heiberger who invested quite some time in testing
uncharted territory and made valuable suggestions for improving this function}
