\name{RcmdrPlugin.mosaic-package}
\alias{RcmdrPlugin.mosaic-package}
\alias{RcmdrPlugin.mosaic}
\docType{package}
\title{
Adds an Rcmdr menu items for manipulating structables and producing mosaic and assoc plot
}
\description{
The menu item opens a dialog bx which allows to change the structure
of structables visually.
}
\details{
\tabular{ll}{
Package: \tab RcmdrPlugin.mosaic\cr
Type: \tab Package\cr
Version: \tab 1.0-1\cr
Date: \tab 2011-07-02\cr
License: \tab GPL 2.0\cr
LazyLoad: \tab yes\cr
}
}
\author{
Erich Neuwirth

Maintainer: Erich Neuwirth <erich.neuwirth@univie.ac.at>
}
\keyword{ graphics,
 hplot}

 
\note{Special thanks go to Richard Heiberger who invested quite some time in testing
uncharted territory and made valuable suggestions for improving the interface to mosaic.}
