\name{RcmdrPlugin.Calvin-internal}
\title{Internal RcmdrPlugin.Calvin objects}
\alias{meansd}
\alias{fvnum}
\alias{favstats}
\alias{xyPlot}
\alias{bwPlot}
\alias{lightBlue}
\alias{darkBlue}
\alias{col.navy}
\alias{col.calvin}
\alias{latticeShowSettings}
\alias{latticeTheme}
\alias{latticeHistogram}
\alias{HmiscStatisticsTable}
\alias{conditionBox}
\description{Internal RcmdrPlugin.Calvin objects.}
\details{These are not intended to be called by the user, but
favstats and meansd could be of some use.}
